Project3/requirements.txt
Flask==3.0.2
tensorflow==2.13.1
Pillow